BM Review and Search (#17)
* Update and Delete Review

* Testing Update and Delete Review

* Deleting unused imports and methods

* Minor fixes

* Changing Request Part for Request Body

* Search Book by title and Author by name

* Requested fixes in PR

* Fixing patterns

* Fixing Specifications and adding Pagination

* Adding search by title to Book

* minor corrections

+++ src/main/java/com/austral/bookin/dto/book/BookDTO.java
@@ -17,5 +17,5 @@ public class BookDTO {
     private String language;
     private Date date;
     private byte[] photo;
-    private double stars;
+    private float stars;
 }
